mcp-client/requirements.txt
mcp==1.9.4

langchain==0.3.26
langchain-core==0.3.67
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langchain-chroma==0.2.4
langchain-mcp-adapters==0.1.7
langgraph>=0.6.4

fastapi==0.115.13
uvicorn==0.34.3

pydantic==2.11.7

httpx==0.28.1

pandas==2.3.0

python-dotenv==1.1.0

typing-extensions==4.12.2

sentence-transformers

langchain-ollama
